language-learning-bot/requirements.txt
# Telegram Bot
python-telegram-bot>=13.13
aiogram>=2.21.0

# Web framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
pymongo>=4.3.3

# Excel processing
openpyxl>=3.1.2
pandas>=2.0.0
xlrd>=2.0.1

# Configuration
python-dotenv>=1.0.0
hydra-core>=1.3.2
omegaconf>=2.3.0

# Async
httpx>=0.24.0
asyncio>=3.4.3

# Logging
loguru>=0.7.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Code quality
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0

# Other
pydub>=0.25.1  # For audio processing

motor>=3.3.0
pymongo>=4.5.0

# Распознавание речи
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.22.0
imageio-ffmpeg>=0.4.8